network: Don't store the network info as pointers if slices used
By storing the network information such as routes as pointers, we
end up having issues if we try to remove some elements from the list
as at the same time this list is parsed through by a loop.

For interfaces, we can use maps instead, because we have a way to
keep track of them based on their names. Maps are safer and allow
to keep pointers in this case.

In both cases, routes and interfaces list can be modified properly
while we loop over every items.

Fixes #226

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>