grpc: Don't close the terminal on CloseStdin() call
CloseStdin() is supposed to close only the standard input, which
helps the process to terminate in some cases. But for the case where
we are using a terminal, it should be closed by the agent once the
WaitProcess() will complete. Otherwise, closing it before the end of
the process could prevent some output from the terminal to reach the
shim.

Fixes #283

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>