agent: fix the issue of missing close process terminal
Kata assumes once a process exits, it's IO will also
close automatically, thus the client will wait on the
process's IO closed and then do some cleanup for this
process. But the issue is that if the process forked
some background children processes running as daemon
process, thus those children process will intherit its
parent terminal, and even its parent exited, its terminal
will not close as the children running.

So this commit will try to fix this issue by introducing
a pipe, and let the IO read epolling on one end of this pipe
and the process's pty master. Once the process exits, it
will close the other end of the pipe to notify the IO read
that the process has exited and and there's no need to wait
on its IO.

Fixes: #370

Signed-off-by: fupan <lifupan@gmail.com>