agent: Fix the issue of stdout hang on builtin proxy
Since the Yamux's keepalive has been disabled both
on the server and the client side, and this brings
a weird issue where the communication between the
proxy and the agent hangs. The same issue has been
fixed in kata proxy by:
"https://github.com/kata-containers/proxy/pull/91".

This commit just cherry-pick that patch here to
fix the same issue on kata builtin proxy.

Fixes: #396

Signed-off-by: fupan <lifupan@gmail.com>