vendor: Update vendor/github.com/containerd/console
Fixes: #410

The console package in vendor contains a big endian bug. The issue has
already been solved in runc. See PRs:
 - https://github.com/opencontainers/runc/pull/1727
 - https://github.com/containerd/console/pull/20

The console version has been updated from commit 84eeaae905 to commit
2748ece166 (commit that fixed the big-endian issue)

The console package has been added in the Gopkg.toml because the console
package is not directly used in the code and the dep command doesn't
automatically update the console package.

Signed-off-by: Alice Frosi <afrosi@de.ibm.com>